- Add mockk library for unit test - Add DefaultUniversityRepositoryTest class for cover function unit test - add safeApiCall for handle error response from API - update NetworkDataSource and DefaultUniversityRepository function - update DefaultUniversityRepositoryTest - update viewModel for bug pullToRefresh on offline mode

+++ app/src/main/java/com/vikination/universitylistapp/data/DefaultUniversityRepository.kt
@@ -1,10 +1,13 @@
 package com.vikination.universitylistapp.data
 
+import com.vikination.universitylistapp.data.UniversityRepository.Resource
 import com.vikination.universitylistapp.data.source.local.UniversityDao
 import com.vikination.universitylistapp.data.source.network.NetworkDataSource
+import com.vikination.universitylistapp.data.source.network.NetworkUniversity
 import com.vikination.universitylistapp.di.DefaultDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
@@ -24,12 +27,19 @@ class DefaultUniversityRepository @Inject constructor(
         }
     }
 
-    override suspend fun getUniversitiesFromNetwork() {
-        withContext(dispatcher){
-            val universityListFromNetwork = networkDataSource.loadAllUniversities()
-            localDataSource.deleteAll()
-            localDataSource.upsertAll(universityListFromNetwork.toLocalUniversityList())
+    override fun getUniversitiesFromNetwork(): Flow<Resource<List<NetworkUniversity>>> {
+        return flow {
+            val response = safeApiCall { networkDataSource.loadAllUniversities() }
+            withContext(dispatcher){
+                if (response is Resource.Success){
+                    val universityListFromNetwork = response.data ?: emptyList()
+                    localDataSource.deleteAll()
+                    localDataSource.upsertAll(universityListFromNetwork.toLocalUniversityList())
+                }
+            }
+            emit(response)
         }
+
     }
 
     /**
